# Fast implementation of digits with base a power of 2