This PR adds a fast specialised implementation of `digits(::Integer, base = 2)`.  @ararslan said that making `base = 2` a special case wouldn't be a bad idea.  Performance with other bases shouldn't be affected.

Before this PR:
```julia
julia> @benchmark digits(3, base = 2)
BenchmarkTools.Trial: 
  memory estimate:  96 bytes
  allocs estimate:  1
  --------------
  minimum time:     40.681 ns (0.00% GC)
  median time:      41.959 ns (0.00% GC)
  mean time:        52.063 ns (14.25% GC)
  maximum time:     43.487 μs (99.85% GC)
  --------------
  samples:          10000
  evals/sample:     991

julia> @benchmark digits(3, base = 10)
BenchmarkTools.Trial: 
  memory estimate:  96 bytes
  allocs estimate:  1
  --------------
  minimum time:     35.241 ns (0.00% GC)
  median time:      36.893 ns (0.00% GC)
  mean time:        47.171 ns (16.41% GC)
  maximum time:     43.818 μs (99.88% GC)
  --------------
  samples:          10000
  evals/sample:     992

julia> @benchmark digits(1234567890123456, base = 2)
BenchmarkTools.Trial: 
  memory estimate:  496 bytes
  allocs estimate:  1
  --------------
  minimum time:     658.308 ns (0.00% GC)
  median time:      667.686 ns (0.00% GC)
  mean time:        730.018 ns (5.45% GC)
  maximum time:     247.337 μs (99.70% GC)
  --------------
  samples:          10000
  evals/sample:     159

julia> @benchmark digits(1234567890123456, base = 10)
BenchmarkTools.Trial: 
  memory estimate:  208 bytes
  allocs estimate:  1
  --------------
  minimum time:     222.505 ns (0.00% GC)
  median time:      225.409 ns (0.00% GC)
  mean time:        247.892 ns (5.90% GC)
  maximum time:     87.951 μs (99.67% GC)
  --------------
  samples:          10000
  evals/sample:     487
```

After this PR:
```julia
julia> @benchmark digits(3, base = 2)
BenchmarkTools.Trial: 
  memory estimate:  96 bytes
  allocs estimate:  1
  --------------
  minimum time:     34.707 ns (0.00% GC)
  median time:      35.394 ns (0.00% GC)
  mean time:        43.532 ns (14.30% GC)
  maximum time:     37.210 μs (99.85% GC)
  --------------
  samples:          10000
  evals/sample:     993

julia> @benchmark digits(3, base = 10)
BenchmarkTools.Trial: 
  memory estimate:  96 bytes
  allocs estimate:  1
  --------------
  minimum time:     34.356 ns (0.00% GC)
  median time:      35.174 ns (0.00% GC)
  mean time:        43.805 ns (14.93% GC)
  maximum time:     38.129 μs (99.86% GC)
  --------------
  samples:          10000
  evals/sample:     993

julia> @benchmark digits(1234567890123456, base = 2)
BenchmarkTools.Trial: 
  memory estimate:  496 bytes
  allocs estimate:  1
  --------------
  minimum time:     75.659 ns (0.00% GC)
  median time:      77.740 ns (0.00% GC)
  mean time:        88.100 ns (8.38% GC)
  maximum time:     30.302 μs (99.61% GC)
  --------------
  samples:          10000
  evals/sample:     971

julia> @benchmark digits(1234567890123456, base = 10)
BenchmarkTools.Trial: 
  memory estimate:  208 bytes
  allocs estimate:  1
  --------------
  minimum time:     215.581 ns (0.00% GC)
  median time:      217.726 ns (0.00% GC)
  mean time:        236.653 ns (5.16% GC)
  maximum time:     73.486 μs (99.63% GC)
  --------------
  samples:          10000
  evals/sample:     525
```