@antoine-levitt 2019-04-15
Have you tried benchmarking inside a function? Can't the compiler do constant propagation by itself here?