@giordano 2019-04-15
Like this?
```julia
julia> function test()
           n = 1234567890123456
           digits(n, base = 2)
       end
test (generic function with 1 method)

julia> @btime test();
  74.073 ns (1 allocation: 496 bytes)

julia> function test()
           n = 1234567890123456
           digits(n, base = 10)
       end
test (generic function with 1 method)

julia> @btime test();
  215.014 ns (1 allocation: 208 bytes)
```