@giordano 2019-04-15
With this change
```diff
@@ -755,9 +755,17 @@ function digits!(a::AbstractVector{T}, n::Integer; base::Integer = 10) where T<:
     isempty(a) && return a
 
     if base > 0
-        for i in eachindex(a)
-            n, d = divrem(n, base)
-            a[i] = d
+        if ispow2(base)
+            k = trailing_zeros(base)
+            c = base - 1
+            for i in eachindex(a)
+                a[i] = (n >> (k * (i - 1))) & c
+            end
+        else
+            for i in eachindex(a)
+                n, d = divrem(n, base)
+                a[i] = d
+            end
         end
     else
         # manually peel one loop iteration for type stability
```
we can have specialised implementation for all bases that are powers of 2.  Here are benchmarks with the proposed change to the PR:
```julia
# Before
julia> @btime digits(1, base = 2);
  33.756 ns (1 allocation: 96 bytes)

julia> @btime digits(1234567890123456, base = 2);
  660.050 ns (1 allocation: 496 bytes)

julia> @btime digits(81985529216486895, base = 2);
  759.219 ns (1 allocation: 544 bytes)

julia> @btime digits(1, base = 16);
  34.774 ns (1 allocation: 96 bytes)

julia> @btime digits(81985529216486895, base = 16);
  219.475 ns (1 allocation: 208 bytes)


# After
julia> @btime digits(1, base = 2);
  33.044 ns (1 allocation: 96 bytes)

julia> @btime digits(1234567890123456, base = 2);
  111.649 ns (1 allocation: 496 bytes)

julia> @btime digits(81985529216486895, base = 2);
  122.859 ns (1 allocation: 544 bytes)

julia> @btime digits(1, base = 16);
  32.785 ns (1 allocation: 96 bytes)

julia> @btime digits(81985529216486895, base = 16);
  58.451 ns (1 allocation: 208 bytes)
```
Note that performance of `base = 2` gets a bit worse compared to the current version of the PR, but still way better than the generic implementation in `master`.