@giordano 2019-04-15
Ok, I've fixed the first index as suggested by Rafael and implemented the fast method for all powers of 2, as well as added more tests.