@StefanKarpinski 2019-04-24
I guess the question is who should make a decision here? This seems like a straightforward speedup. And which version should we go with?