@giordano 2019-04-24
> who should make a decision here?

I have no idea :joy: 

> And which version should we go with?

I'd suggest the last one with any power of 2, I think it's a common case to see the expansion in base 8 and 16, not only base 2.  Anyway, I can easily revert to the previous version if we want to prioritise `base = 2`.