@stevengj 2019-04-25
Should there be a faster `ndigits` in this case too, or does it spend a negligible amount of time in `ndigits`?