@giordano 2019-04-25
It doesn't seem that `ndigits` has a significant impact on performance of `digits`:
```julia
julia> Profile.clear(); @profile for n in 1:1_000_000
           digits(1234567890123456, base = 2)
       end

julia> Profile.print()
133 ./task.jl:268; (::getfield(REPL, Symbol("##26#27")){REPL.REPLBackend})()
 133 ...e/julia/stdlib/v1.3/REPL/src/REPL.jl:118; macro expansion
  133 ...e/julia/stdlib/v1.3/REPL/src/REPL.jl:86; eval_user_input(::Any, ::REPL.REPLBackend)
   133 ./boot.jl:330; eval(::Module, ::Any)
    1   ./intfuncs.jl:755; #digits!#326(::Int64, ::typeof(digits!), ::Array{Int64...
    1   ./intfuncs.jl:513; ndigits0z(::Int64, ::Int64)
    131 .../stdlib/v1.3/Profile/src/Profile.jl:25; top-level scope
     131 ./REPL[17]:2; macro expansion
      131 ./none:0; #digits
       131 ./intfuncs.jl:710; #digits#324
        131 ./none:0; #digits
         131 ./intfuncs.jl:714; #digits#325
          34 ./array.jl:441; zeros
           26 ./array.jl:445; zeros
            26 ./boot.jl:413; Type
             26 ./boot.jl:404; Type
           8  ./array.jl:446; zeros
            8 ./array.jl:295; fill!(::Array{Int64,1}, ::Int64)
             7 ./array.jl:766; setindex!
             1 ./range.jl:595; iterate
              1 ./promotion.jl:403; ==
          93 ./none:0; #digits!
           3  ./int.jl:0; #digits!#326(::Int64, ::typeof(digits!), ::Array{In...
           9  ./int.jl:444; #digits!#326(::Int64, ::typeof(digits!), ::Array{In...
           3  ./intfuncs.jl:0; #digits!#326(::Int64, ::typeof(digits!), ::Array{In...
           1  ./intfuncs.jl:752; #digits!#326(::Int64, ::typeof(digits!), ::Array{In...
           2  ./intfuncs.jl:755; #digits!#326(::Int64, ::typeof(digits!), ::Array{In...
           75 ./intfuncs.jl:762; #digits!#326(::Int64, ::typeof(digits!), ::Array{In...
            15 ./array.jl:766; setindex!
            10 ./int.jl:293; &
            10 ./int.jl:437; >>
            38 ./int.jl:444; >>
             4  ./array.jl:766; <<
             27 ./int.jl:439; <<
             7  ./int.jl:424; <=
          4  ./none:0; #ndigits
           4 ./intfuncs.jl:544; #ndigits#322
            3 ./intfuncs.jl:513; ndigits0z(::Int64, ::Int64)
             1 ./int.jl:49; <
            1 ./intfuncs.jl:516; ndigits0z(::Int64, ::Int64)
```
Actually, `ndigits` is already optimised for `base` = 2, 8, 16: https://github.com/JuliaLang/julia/blob/93c9ae488d48d14214873d683d8f039313bdbd4d/base/intfuncs.jl#L459-L461