@giordano 2019-04-25
I've changed the condition for the fast path to
```julia
ispow2(base) && n >= 0 && n isa Base.BitInteger
```
`n` must be non-negative, otherwise the result would be simply wrong (and I've added tests for that), I'm also testing that `n` is a `BitInteger` because I think that in general it's not guaranteed that we can do bit twiddling for non bit-types.