@giordano 2019-04-25
How about
```diff
@@ -460,6 +460,10 @@ function ndigits0zpb(x::Integer, b::Integer)
         b == 8  && return (sizeof(x)<<3 - leading_zeros(x) + 2) ÷ 3
         b == 16 && return sizeof(x)<<1 - leading_zeros(x)>>2
         b == 10 && return bit_ndigits0z(x)
+        if ispow2(b)
+            dv, rm = divrem(sizeof(x)<<3 - leading_zeros(x), trailing_zeros(b))
+            return iszero(rm) ? dv : dv + 1
+        end
     end
 
     d = 0
```
in `ndigits0zpb`?

Before this change:
```julia
julia> @btime ndigits(81985529216486895, base = 4);
  29.934 ns (0 allocations: 0 bytes)

julia> @btime digits(81985529216486895, base = 4);
  106.863 ns (1 allocation: 336 bytes)
```
After it:
```julia
julia> @btime ndigits(81985529216486895, base = 4);
  8.963 ns (0 allocations: 0 bytes)

julia> @btime digits(81985529216486895, base = 4);
  85.248 ns (1 allocation: 336 bytes)
```
I think it'd be nicer to remove the special cases for 2, 8, 16, but this generic method I came up with is a bit slower than those super-optimised cases.