@giordano 2019-05-02
Any comment on the change to `ndigits0zpb` proposed above?