@StefanKarpinski 2019-05-02
That seems good to me. In general, as long as it's a) correct, b) faster, and c) not too insane, I'm happy with any of these changes.