@giordano 2019-05-02
Ok, I installed the proposed change.  Here are some benchmarks:

before the PR:
```julia
julia> for base in (2, 3, 4, 5, 8, 10, 16),
           n in (1, 3, 1234567890123456, 81985529216486895)
           println("n = $n, base = $base")
           @btime digits($n, base = $base)
       end
n = 1, base = 2
  33.569 ns (1 allocation: 96 bytes)
n = 3, base = 2
  38.626 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 2
  642.904 ns (1 allocation: 496 bytes)
n = 81985529216486895, base = 2
  759.684 ns (1 allocation: 544 bytes)
n = 1, base = 3
  37.212 ns (1 allocation: 96 bytes)
n = 3, base = 3
  44.564 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 3
  432.618 ns (1 allocation: 336 bytes)
n = 81985529216486895, base = 3
  509.099 ns (1 allocation: 368 bytes)
n = 1, base = 4
  36.202 ns (1 allocation: 96 bytes)
n = 3, base = 4
  37.109 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 4
  341.088 ns (1 allocation: 288 bytes)
n = 81985529216486895, base = 4
  399.398 ns (1 allocation: 336 bytes)
n = 1, base = 5
  37.231 ns (1 allocation: 96 bytes)
n = 3, base = 5
  37.755 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 5
  307.689 ns (1 allocation: 256 bytes)
n = 81985529216486895, base = 5
  362.034 ns (1 allocation: 288 bytes)
n = 1, base = 8
  33.838 ns (1 allocation: 96 bytes)
n = 3, base = 8
  33.675 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 8
  232.650 ns (1 allocation: 224 bytes)
n = 81985529216486895, base = 8
  273.321 ns (1 allocation: 240 bytes)
n = 1, base = 10
  34.165 ns (1 allocation: 96 bytes)
n = 3, base = 10
  34.032 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 10
  218.868 ns (1 allocation: 208 bytes)
n = 81985529216486895, base = 10
  246.200 ns (1 allocation: 224 bytes)
n = 1, base = 16
  33.623 ns (1 allocation: 96 bytes)
n = 3, base = 16
  33.525 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 16
  183.280 ns (1 allocation: 192 bytes)
n = 81985529216486895, base = 16
  221.509 ns (1 allocation: 208 bytes)
```
After the PR:
```julia
julia> for base in (2, 3, 4, 5, 8, 10, 16),
           n in (1, 3, 1234567890123456, 81985529216486895)
           println("n = $n, base = $base")
           @btime digits($n, base = $base)
       end
n = 1, base = 2
  33.610 ns (1 allocation: 96 bytes)
n = 3, base = 2
  34.253 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 2
  110.220 ns (1 allocation: 496 bytes)
n = 81985529216486895, base = 2
  123.984 ns (1 allocation: 544 bytes)
n = 1, base = 3
  37.236 ns (1 allocation: 96 bytes)
n = 3, base = 3
  44.465 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 3
  430.688 ns (1 allocation: 336 bytes)
n = 81985529216486895, base = 3
  509.145 ns (1 allocation: 368 bytes)
n = 1, base = 4
  34.196 ns (1 allocation: 96 bytes)
n = 3, base = 4
  34.354 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 4
  81.159 ns (1 allocation: 288 bytes)
n = 81985529216486895, base = 4
  86.555 ns (1 allocation: 336 bytes)
n = 1, base = 5
  37.245 ns (1 allocation: 96 bytes)
n = 3, base = 5
  37.056 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 5
  306.630 ns (1 allocation: 256 bytes)
n = 81985529216486895, base = 5
  359.809 ns (1 allocation: 288 bytes)
n = 1, base = 8
  33.450 ns (1 allocation: 96 bytes)
n = 3, base = 8
  33.501 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 8
  59.394 ns (1 allocation: 224 bytes)
n = 81985529216486895, base = 8
  61.872 ns (1 allocation: 240 bytes)
n = 1, base = 10
  34.125 ns (1 allocation: 96 bytes)
n = 3, base = 10
  35.073 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 10
  217.121 ns (1 allocation: 208 bytes)
n = 81985529216486895, base = 10
  244.356 ns (1 allocation: 224 bytes)
n = 1, base = 16
  33.395 ns (1 allocation: 96 bytes)
n = 3, base = 16
  33.392 ns (1 allocation: 96 bytes)
n = 1234567890123456, base = 16
  54.078 ns (1 allocation: 192 bytes)
n = 81985529216486895, base = 16
  57.518 ns (1 allocation: 208 bytes)
```
I also included some benchmark for values bases not powers of 2 (3, 5, 10), to make sure that performance is effectively unaffected.